app/pyproject.toml
[tool.poetry]
name = "wym"
version = "0.1.0"
description = "flask webapp"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
Flask = "^2.2.2"
SQLAlchemy = "^1.4.41"
Jinja2 = "^3.1.2"
pandas = "^1.4.4"
numpy = "^1.23.3"
requests = "^2.28.1"
psycopg2-binary = "^2.9.3"
Flask-SQLAlchemy = "^2.5.1"
SQLAlchemy-Utils = "^0.38.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
